contracts: organize test utils files
