+++ apps/contracts/test/utils/models/entry-point/EntryPointDeployer.ts
@@ -1,5 +1,6 @@
+import { deploy } from '../../helpers/contracts'
+
 import EntryPoint from './EntryPoint'
-import { deploy } from '../../helpers/contracts';
 
 const EntryPointDeployer = {
   async deploy(): Promise<EntryPoint> {
